Check for `da:install_status:...` when restarting
When the server restarts for some reason, it saves the background
task id under `da:restart_status:...`, or as `da:install_status:...`
if a package is being installed to the server.

When someone calls `/api/restart_status`, DA will check if the
given task id key is either `restart_status` or `install_status`,
instead of just checking `restart_status`.